test(core): migrate PaneToolbarTests (Wave D)
9 tests, 12 failure sites preserved 1:1.
